move disk dependencies
