+++ src/analysis/modal-utils.ts
@@ -5,7 +5,8 @@ import {
   getImageUrlFromRuntime,
 } from 'src/analysis/utils/runtime-utils';
 import { getToolLabelFromCloudEnv } from 'src/analysis/utils/tool-utils';
-import { DiskType, GoogleDiskType, PersistentDisk } from 'src/libs/ajax/leonardo/models/disk-models';
+import { DiskType, GoogleDiskType } from 'src/libs/ajax/leonardo/Disks';
+import { PersistentDisk } from 'src/libs/ajax/leonardo/providers/LeoDiskProvider';
 
 export interface IComputeConfig {
   masterMachineType: string;
