+++ src/analysis/modals/ComputeModal/AzureComputeModal/AzurePersistentDiskSection.ts
@@ -4,7 +4,8 @@ import { AboutPersistentDiskSection } from 'src/analysis/modals/ComputeModal/Abo
 import { AzurePersistentDiskSizeSelectInput } from 'src/analysis/modals/ComputeModal/AzureComputeModal/AzurePersistentDiskSizeSelectInput';
 import { PersistentDiskTypeInputContainer } from 'src/analysis/modals/ComputeModal/PersistentDiskTypeInputContainer';
 import { computeStyles } from 'src/analysis/modals/modalStyles';
-import { AzureDiskType, AzurePdType, AzurePersistentDiskOptions } from 'src/libs/ajax/leonardo/models/disk-models';
+import { AzureDiskType } from 'src/libs/ajax/leonardo/Disks';
+import { AzurePdType } from 'src/libs/ajax/leonardo/providers/LeoDiskProvider';
 import { defaultAzurePersistentDiskType } from 'src/libs/azure-utils';
 
 export interface AzurePersistentDiskSectionProps {
@@ -16,6 +17,20 @@ export interface AzurePersistentDiskSectionProps {
   onChangePersistentDiskSize: (size: number | null | undefined) => void;
   onClickAbout: () => void;
 }
+
+export const AzurePersistentDiskOptions: AzurePdType[] = [
+  {
+    value: 'Standard_LRS',
+    label: 'Standard HDD',
+  },
+  // TODO: Disabled the SSD option and the test in
+  // AzurePersistentDiskInputTest test until SSD is properly implemented and tested.
+  // Main blocker: Cost calculation.
+  // {
+  //   value: 'StandardSSD_LRS',
+  //   label: 'Standard SSD',
+  // },
+];
 
 export const AzurePersistentDiskSection = (props: AzurePersistentDiskSectionProps): ReactNode => {
   const {
