+++ src/analysis/modals/ComputeModal/GcpComputeModal/GcpComputeModal.test.ts
@@ -35,10 +35,9 @@ import {
 } from 'src/analysis/utils/runtime-utils';
 import { runtimeToolLabels, runtimeTools, terraSupportedRuntimeImageIds } from 'src/analysis/utils/tool-utils';
 import { Ajax } from 'src/libs/ajax';
-import { PersistentDisk } from 'src/libs/ajax/leonardo/models/disk-models';
 import { cloudServiceTypes, NormalizedComputeRegion } from 'src/libs/ajax/leonardo/models/runtime-config-models';
 import { runtimeStatuses } from 'src/libs/ajax/leonardo/models/runtime-models';
-import { leoDiskProvider } from 'src/libs/ajax/leonardo/providers/LeoDiskProvider';
+import { LeoDiskProvider, leoDiskProvider, PersistentDisk } from 'src/libs/ajax/leonardo/providers/LeoDiskProvider';
 import { RuntimeAjaxContractV1, RuntimesAjaxContract } from 'src/libs/ajax/leonardo/Runtimes';
 import { formatUSD } from 'src/libs/utils';
 import { asMockedFn, renderWithAppContexts as render } from 'src/testing/test-utils';
@@ -51,7 +50,18 @@ jest.mock('src/libs/notifications', () => ({
   },
 }));
 
-jest.mock('src/libs/ajax/leonardo/providers/LeoDiskProvider');
+type DiskProviderExports = typeof import('src/libs/ajax/leonardo/providers/LeoDiskProvider');
+jest.mock('src/libs/ajax/leonardo/providers/LeoDiskProvider', (): DiskProviderExports => {
+  return {
+    ...jest.requireActual('src/libs/ajax/leonardo/providers/LeoDiskProvider'),
+    leoDiskProvider: {
+      details: jest.fn(),
+      update: jest.fn(),
+      delete: jest.fn(),
+    } as Partial<LeoDiskProvider> as LeoDiskProvider,
+  };
+});
+
 jest.mock('src/libs/ajax');
 jest.mock('src/analysis/utils/cost-utils');
 jest.mock('src/libs/config', () => ({
