+++ src/analysis/modals/ComputeModal/PersistentDiskTypeInput.test.ts
@@ -5,7 +5,8 @@ import {
   PersistentDiskTypeInput,
   PersistentDiskTypeInputProps,
 } from 'src/analysis/modals/ComputeModal/PersistentDiskTypeInput';
-import { GoogleDiskType, GooglePdType } from 'src/libs/ajax/leonardo/models/disk-models';
+import { GoogleDiskType } from 'src/libs/ajax/leonardo/Disks';
+import { GooglePdType } from 'src/libs/ajax/leonardo/providers/LeoDiskProvider';
 import { renderWithAppContexts as render } from 'src/testing/test-utils';
 
 const defaultPersistentDiskTypeInputProps: PersistentDiskTypeInputProps<GoogleDiskType, GooglePdType> = {
