+++ src/libs/ajax/leonardo/models/core-models.ts
@@ -1,6 +1,3 @@
-import { App } from 'src/libs/ajax/leonardo/models/app-models';
-import { PersistentDisk } from 'src/libs/ajax/leonardo/models/disk-models';
-import { Runtime } from 'src/libs/ajax/leonardo/models/runtime-models';
 import { CloudProvider } from 'src/workspaces/utils';
 
 export interface AuditInfo {
@@ -21,5 +18,3 @@ export interface CloudContext {
   cloudProvider: CloudProvider;
   cloudResource: string;
 }
-
-export type Resource = PersistentDisk | App | Runtime;
