+++ src/pages/EnvironmentsPage/environmentsPermissions.ts
@@ -2,8 +2,8 @@ import { LeoResourcePermissionsProvider } from 'src/analysis/Environments/Enviro
 import { getCreatorForCompute, isResourceDeletable } from 'src/analysis/utils/resource-utils';
 import { cromwellAppToolLabels } from 'src/analysis/utils/tool-utils';
 import { App } from 'src/libs/ajax/leonardo/models/app-models';
-import { PersistentDisk } from 'src/libs/ajax/leonardo/models/disk-models';
 import { ListRuntimeItem, Runtime } from 'src/libs/ajax/leonardo/models/runtime-models';
+import { PersistentDisk } from 'src/libs/ajax/leonardo/providers/LeoDiskProvider';
 import { getTerraUser } from 'src/libs/state';
 
 const makePermissionsProvider = (userEmailGetter: () => string): LeoResourcePermissionsProvider => {
